✨(core) separate graphs as apps
As we cannot use Flask blueprints to separate graphs from a main
application, we decided to stick to Dash guidelines to serve multiple
apps from the same wsgi instance.

We now have a Dash application per graph and a main Flask app, all
routed via a dedicated middleware.
