+++ src/potsie/server/__main__.py
@@ -1,0 +1,6 @@
+from werkzeug.serving import run_simple
+
+from .wsgi import app
+
+
+run_simple("0.0.0.0", 8050, app, use_reloader=True, use_debugger=True)
