+++ src/potsie/server/config.py
@@ -1,0 +1,1 @@
+external_stylesheets = ["https://codepen.io/chriddyp/pen/bWLwgP.css"]
