+++ src/potsie/server/video/config.py
@@ -1,0 +1,1 @@
+video_apps_pathname_prefix = "/video"
